Conflicts:
	DesignerRHN/DesignerRHN-war/nbproject/faces-config.NavData

+++ DesignerRHN/DesignerRHN-ejb/src/java/InterfaceAdministrar/AdministrarMotivosCambiosSueldosInterface.java
@@ -26,5 +26,5 @@ public interface AdministrarMotivosCambiosSueldosInterface {
 
     public MotivosCambiosSueldos mostrarMotivoCambioCargo(BigInteger secMotivosCambiosSueldos);
 
-    public Long verificarBorradoVS(BigInteger secuenciaMovitoCambioSueldo);
+    public BigInteger verificarBorradoVS(BigInteger secuenciaMovitoCambioSueldo);
 }
